add folder sharing

+++ server/src/db/file.queries.ts
@@ -1,3 +1,4 @@
+import { File } from "@prisma/client";
 import { prisma } from "./queries";
 
 export const getFilesByUserId = async (userId: number, type: string) => {
@@ -64,11 +65,29 @@ export const getFileUrl = async (fileId: number) => {
 
 export const deleteFile = async (fileId: number) => {
   try {
-    const file = await prisma.file.delete({
-      where: { id: fileId },
+    return await prisma.$transaction(async (prisma) => {
+      const file = await prisma.file.findUnique({
+        where: { id: fileId },
+        select: { folderId: true, size: true, url: true },
+      });
+      if (!file) throw new Error("File not found");
+
+      await prisma.file.delete({
+        where: { id: fileId },
+      });
+
+      if (file.folderId) {
+        await prisma.folder.update({
+          where: { id: file.folderId },
+          data: {
+            size: { decrement: file.size },
+          },
+        });
+      }
+
+      console.log(`File with ID ${fileId} deleted successfully.`);
+      return file.url;
     });
-    console.log(`File with ID ${fileId} deleted successfully.`);
-    return file;
   } catch (error) {
     console.error("Error deleting file:", error);
     throw new Error("Failed to delete file");
@@ -77,11 +96,53 @@ export const deleteFile = async (fileId: number) => {
 
 export const updateFile = async (fileId: number, data: File) => {
   try {
-    await prisma.file.update({
-      where: { id: fileId },
-      data: data,
+    return await prisma.$transaction(async (prisma) => {
+      const oldFile = await prisma.file.findUnique({
+        where: { id: fileId },
+      });
+      if (!oldFile) throw new Error("File not found");
+
+      const updatedFile = await prisma.file.update({
+        where: { id: fileId },
+        data: data,
+      });
+
+      // Update folder sizes if folderId changes
+      if (data.folderId && data.folderId !== oldFile.folderId) {
+        // Decrement size from the old folder
+        if (oldFile.folderId) {
+          await prisma.folder.update({
+            where: { id: oldFile.folderId },
+            data: {
+              size: { decrement: oldFile.size },
+            },
+          });
+        }
+      }
+
+      // Increment size in the new folder
+      if (data.folderId) {
+        await prisma.folder.update({
+          where: { id: data.folderId },
+          data: {
+            size: { increment: data.size ?? oldFile.size },
+          },
+        });
+      }
+
+      // Update folder size if only file size changes
+      if (data.size && data.size !== oldFile.size) {
+        if (oldFile.folderId) {
+          await prisma.folder.update({
+            where: { id: oldFile.folderId },
+            data: {
+              size: { increment: data.size - oldFile.size },
+            },
+          });
+        }
+      }
+      console.log(`File with ID ${fileId} updated successfully.`);
     });
-    console.log(`File with ID ${fileId} updated successfully.`);
   } catch (error) {
     console.error("Error updating file:", error);
     throw new Error("Failed to update file");
@@ -97,17 +158,28 @@ export const createFile = async (
   if (!file) return;
   const { originalname, size } = file;
   try {
-    const file = await prisma.file.create({
-      data: {
-        name: originalname,
-        size,
-        url: fileUrl,
-        userId,
-        folderId,
-      },
+    return await prisma.$transaction(async (prisma) => {
+      const file = await prisma.file.create({
+        data: {
+          name: originalname,
+          size,
+          url: fileUrl,
+          userId,
+          folderId,
+        },
+      });
+
+      if (folderId) {
+        await prisma.folder.update({
+          where: { id: folderId },
+          data: {
+            size: { increment: size },
+          },
+        });
+      }
+      console.log(`File with ID ${file.id} created successfully.`);
+      return file;
     });
-    console.log(`File with ID ${file.id} created successfully.`);
-    return file;
   } catch (error) {
     console.error("Error creating file:", error);
     throw new Error("Failed to create file");
